+++ libMx/fh_autoscale.h
@@ -7,7 +7,7 @@
  *
  *---------------------------------------------------------------------------
  *
- * Copyright 2001 Illinois Institute of Technology
+ * Copyright 2001, 2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -21,7 +21,7 @@ typedef struct {
 	MX_RECORD *autoscale_scaler_record;
 	MX_RECORD *autoscale_record;
 
-	int change_request;
+	long change_request;
 } MX_AUTOSCALE_MEASUREMENT_FAULT;
 
 MX_API mx_status_type mxfh_autoscale_create_handler(
